@@ -4,7 +4,7 @@ version: 1.15.0+40
 homepage: https://github.com/amake/orgro
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   dynamic_fonts: ^2.0.0
