@@ -2,11 +2,9 @@ import 'package:flutter/material.dart';
 
 PopupMenuItem<VoidCallback> readerModeMenuItem(
   BuildContext context, {
-  @required bool enabled,
-  @required Function(bool) onChanged,
+  required bool enabled,
+  required Function(bool) onChanged,
 }) {
-  assert(enabled != null);
-  assert(onChanged != null);
   return CheckedPopupMenuItem<VoidCallback>(
     checked: enabled,
     value: () => onChanged(!enabled),
@@ -16,11 +14,10 @@ PopupMenuItem<VoidCallback> readerModeMenuItem(
 
 class ReaderModeButton extends StatelessWidget {
   const ReaderModeButton({
-    @required this.enabled,
-    @required this.onChanged,
-    Key key,
-  })  : assert(enabled != null),
-        super(key: key);
+    required this.enabled,
+    required this.onChanged,
+    Key? key,
+  }) : super(key: key);
 
   final bool enabled;
   final Function(bool) onChanged;
